add test unit for rest controller

+++ src/main/java/com/test/calculator/Calculator.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class Calculator {
-    public long sum(long a , long b){
+    public int sum(int a , int b){
         return  a+b;
     }
 }
